// packages/react/src/lib/components/secondary-navigation-level-two/secondary-navigation-level-two.d.ts
import * as React from 'react';
export declare const PdsSecondaryNavigationLevelTwo: React.FunctionComponent<Partial<{
    title: string;
    href: string;
    activePage: boolean;
    activePageCallback: Function | undefined;
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;